move ExportVectorDataChip to chips module
