+++ src/modules/chips/components/assistChips/ExportVectorDataChip.js
@@ -4,7 +4,7 @@
 import { html } from 'lit-html';
 import { $injector } from '../../../../injection';
 import { openModal } from '../../../../store/modal/modal.action';
-import { AbstractAssistChip } from '../../../chips/components/assistChips/AbstractAssistChip';
+import { AbstractAssistChip } from './AbstractAssistChip';
 import exportSvg from './assets/download.svg';
 
 const Update_Data = 'update_data';
